alts: convert TsiHandshakeHandler to fit new ProtocolNegotiation style (#6104)
